Add sprite (no animation) and decrease camera size
We added a sprite to the object to see how we need to adjust the size of certain components to make it feel more real. We also decreased the camera size to accommodate the size of the character. Decreasing the size of the camera meant that things felt faster, so we made some adjustments to the numbers that control the speed.