Add ground checks and friction to the Player
We added another collider that is only at the feet of the player. We also added layers to the player and walls and added a layer mask to the player so we can monitor the interaction between these layers. For friction, we just multiple the velocity by a number less than 1.